+++ TexiWithMe/src/main/java/com/example/texiwithme/domain/user/controller/UserController.java
@@ -3,9 +3,9 @@ package com.example.texiwithme.domain.user.controller;
 import com.example.texiwithme.domain.user.dto.request.LoginRequestDto;
 import com.example.texiwithme.domain.user.dto.request.SignupRequestDto;
 import com.example.texiwithme.domain.user.dto.respond.LoginRespondDto;
-import com.example.texiwithme.domain.user.service.CreateUserService;
-import com.example.texiwithme.domain.user.service.DeleteUserService;
-import com.example.texiwithme.domain.user.service.LoginService;
+import com.example.texiwithme.domain.user.usecase.CreateUserUseCase;
+import com.example.texiwithme.domain.user.usecase.DeleteUserUseCase;
+import com.example.texiwithme.domain.user.usecase.LoginUseCase;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.security.core.Authentication;
@@ -15,24 +15,24 @@ import org.springframework.web.bind.annotation.*;
 @RequiredArgsConstructor
 public class UserController {
 
-    private final CreateUserService createUserService;
-    private final LoginService loginService;
-    private final DeleteUserService deleteUserService;
+    private final CreateUserUseCase createUserUseCase;
+    private final LoginUseCase loginUseCase;
+    private final DeleteUserUseCase deleteUserUseCase;
 
     @PostMapping("/user/signup")
     @ResponseStatus(HttpStatus.CREATED)
     public void signup (@RequestBody SignupRequestDto signupRequestDto) {
-        createUserService.create(signupRequestDto);
+        createUserUseCase.create(signupRequestDto);
     }
 
     @PostMapping("/user/login")
     public LoginRespondDto login (@RequestBody LoginRequestDto loginRequestDto) {
-        return loginService.login(loginRequestDto);
+        return loginUseCase.login(loginRequestDto);
     }
 
     @DeleteMapping("/user")
     @ResponseStatus(HttpStatus.NO_CONTENT)
     public void deleteAcc (Authentication authentication) {
-        deleteUserService.deleteUser(authentication);
+        deleteUserUseCase.deleteUser(authentication);
     }
 }
