+++ TexiWithMe/src/main/java/com/example/texiwithme/domain/user/usecase/CreateUserUseCase.java
@@ -1,7 +1,7 @@
-package com.example.texiwithme.domain.user.service;
+package com.example.texiwithme.domain.user.usecase;
 
 import com.example.texiwithme.domain.user.dto.request.SignupRequestDto;
 
-public interface CreateUserService {
+public interface CreateUserUseCase {
     void create(SignupRequestDto signupRequestDto);
 }
