+++ TexiWithMe/src/main/java/com/example/texiwithme/domain/user/usecase/DeleteUserUseCase.java
@@ -1,7 +1,7 @@
-package com.example.texiwithme.domain.user.service;
+package com.example.texiwithme.domain.user.usecase;
 
 import org.springframework.security.core.Authentication;
 
-public interface DeleteUserService {
+public interface DeleteUserUseCase {
     void deleteUser (Authentication authentication);
 }
