+++ TexiWithMe/src/main/java/com/example/texiwithme/domain/user/usecase/LoginUseCase.java
@@ -1,8 +1,8 @@
-package com.example.texiwithme.domain.user.service;
+package com.example.texiwithme.domain.user.usecase;
 
 import com.example.texiwithme.domain.user.dto.request.LoginRequestDto;
 import com.example.texiwithme.domain.user.dto.respond.LoginRespondDto;
 
-public interface LoginService {
+public interface LoginUseCase {
     LoginRespondDto login(LoginRequestDto loginRequestDto);
 }
